fix(app-router): keep popstate restoration tests unit-loadable
The popstate scroll race tests imported app-browser-entry directly, which pulls virtual modules that the unit runner cannot load outside the browser entry build graph. That made the PR fail before the new regression tests could run.

Extract the popstate restoration coordinator into a normal helper module, wire the browser entry through it, and keep the tests' fake RSC navigation promises identical to the pending marker they assert against.
